Various minor changes, targets still don't work properly
I made some changes to things like where the player starts, the ground hitbox, the size of the gun and dummy rounds, as well as trying a lot of different things with the targets to get them to work correctly. currently the test target is set for 'On Collision' rather than 'On Trigger' and that seems to be working.